+++ src/lib.rs
@@ -36,7 +36,7 @@
 //!
 //! # Writing Tests
 //!
-//! ```no_run
+//! ```
 //! use insta::assert_debug_snapshot;
 //!
 //! #[test]
@@ -47,7 +47,7 @@
 //!
 //! The recommended flow is to run the tests once, have them fail and check
 //! if the result is okay.  By default the new snapshots are stored next
-//! to the old ones with the extra `.new` extension.  Once you are satisifed
+//! to the old ones with the extra `.new` extension.  Once you are satisfied
 //! move the new files over.  To simplify this workflow you can use
 //! `cargo insta review` which will let you interactively review them:
 //!