+++ CHANGELOG.md
@@ -40,7 +40,7 @@ All notable changes to insta and cargo-insta are documented here.
 **Upgrade Notes:**
 
 Insta used to detect the current test name by using the current thread name. This
-appeared to work well but unfortunatley ran into various limitations. In particular
+appeared to work well but unfortunately ran into various limitations. In particular
 in some cases the thread name was truncated, missing or did not point to the current
 test name. To better support different platforms and situations insta now uses the
 function name instead.