+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to insta and cargo-insta are documented here.
 
-## 0.16.0
+## 1.16.0
 
 - Added `--no-quiet`/`-Q` flag to `cargo insta test` to suppress the
   quiet flag. This works around limitations with custom test harnesses